Introduces a new InterpolatedECDF struct.
This is created by calling a new `interpolate()` method on a regular ECDF. For now it uses linear interpolation, but this may be expanded in the future to support other forms of interpolation between samples. It is possible to merge two interpolated ECDFs, and also calculate the area between them.